Implements tile caching and WMS proxy
This commit introduces a tile caching mechanism to improve performance and reduce the load on the WMS server.

It also adds a WMS proxy to handle requests for map tiles.
The proxy caches tiles locally and serves them to clients,
reducing the need to fetch tiles from the WMS server repeatedly.
The cache can be configured in the config file.

Also includes dependency injection using "samber/do/v2"